+++ android/app/src/main/java/io/metamask/MainApplication.java
@@ -1,7 +1,6 @@
 package io.metamask;
 
 import com.facebook.react.ReactApplication;
-import com.reactnativecommunity.webviewforked.RNCWebViewForkedPackage;
 import com.cmcewen.blurview.BlurViewPackage;
 import android.content.Context;
 import com.facebook.react.PackageList;
@@ -17,6 +16,7 @@ import cl.json.ShareApplication;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 import io.metamask.nativeModules.PreventScreenshotPackage;
+import android.webkit.WebView;
 
 import androidx.multidex.MultiDexApplication;
 
@@ -55,6 +55,8 @@ public class MainApplication extends MultiDexApplication implements ShareApplica
 	@Override
 	public void onCreate() {
 		super.onCreate();
+    if (BuildConfig.DEBUG)
+    { WebView.setWebContentsDebuggingEnabled(true); }
 		SoLoader.init(this, /* native exopackage */ false);
 
 		initializeFlipper(this, getReactNativeHost().getReactInstanceManager());