+++ app/components/UI/Navbar/index.js
@@ -469,15 +469,16 @@ export function getBrowserViewNavbarOptions(navigation) {
 
 	const isHomepage = url => getHost(url) === getHost(HOMEPAGE_URL);
 	const error = navigation.getParam('error', '');
+	const icon = navigation.getParam('icon', null);
 
 	if (url && !isHomepage(url)) {
 		isHttps = url && url.toLowerCase().substr(0, 6) === 'https:';
 		const urlObj = new URL(url);
-		hostname = urlObj.hostname.toLowerCase().replace(/^www./, '');
+		hostname = urlObj.hostname.toLowerCase().replace(/^www\./, '');
 		if (isGatewayUrl(urlObj) && url.search(`${AppConstants.IPFS_OVERRIDE_PARAM}=false`) === -1) {
 			const ensUrl = navigation.getParam('currentEnsName', '');
 			if (ensUrl) {
-				hostname = ensUrl.toLowerCase().replace(/^www./, '');
+				hostname = ensUrl.toLowerCase().replace(/^www\./, '');
 			}
 		}
 	} else {
@@ -501,7 +502,14 @@ export function getBrowserViewNavbarOptions(navigation) {
 			</TouchableOpacity>
 		),
 		headerTitle: (
-			<NavbarBrowserTitle error={!!error} navigation={navigation} url={url} hostname={hostname} https={isHttps} />
+			<NavbarBrowserTitle
+				error={!!error}
+				icon={url && !isHomepage(url) ? icon : null}
+				navigation={navigation}
+				url={url}
+				hostname={hostname}
+				https={isHttps}
+			/>
 		),
 		headerRight: (
 			<View style={styles.browserRightButton}>
@@ -679,10 +687,8 @@ export function getOfflineModalNavbar(navigation) {
  * @returns {Object} - Corresponding navbar options containing headerTitle, headerTitle and headerTitle
  */
 export function getWalletNavbarOptions(title, navigation) {
-	const onScanSuccess = data => {
-		if (data.target_address) {
-			navigation.navigate('SendView', { txMeta: data });
-		} else if (data.private_key) {
+	const onScanSuccess = (data, content) => {
+		if (data.private_key) {
 			Alert.alert(
 				strings('wallet.private_key_detected'),
 				strings('wallet.do_you_want_to_import_this_account'),
@@ -715,13 +721,9 @@ export function getWalletNavbarOptions(title, navigation) {
 			}, 500);
 		} else if (data.seed) {
 			Alert.alert(strings('wallet.error'), strings('wallet.logout_to_import_seed'));
-		} else if (data && data.indexOf(AppConstants.MM_UNIVERSAL_LINK_HOST) !== -1) {
+		} else {
 			setTimeout(() => {
-				DeeplinkManager.parse(data);
-			}, 500);
-		} else if ((data && data.indexOf('https://') !== -1) || data.indexOf('http://')) {
-			setTimeout(() => {
-				DeeplinkManager.parse(data);
+				DeeplinkManager.parse(content, { origin: AppConstants.DEEPLINKS.ORIGIN_QR_CODE });
 			}, 500);
 		}
 	};