+++ app/components/Views/ApproveView/Approve/index.js
@@ -37,6 +37,8 @@ import IonicIcon from 'react-native-vector-icons/Ionicons';
 import TransactionReviewDetailsCard from '../../../UI/TransactionReview/TransactionReivewDetailsCard';
 import StyledButton from '../../../UI/StyledButton';
 import currencySymbols from '../../../../util/currency-symbols.json';
+import Logger from '../../../../util/Logger';
+import AppConstants from '../../../../core/AppConstants';
 
 const { BNToHex, hexToBN } = util;
 const styles = StyleSheet.create({
@@ -212,6 +214,8 @@ const styles = StyleSheet.create({
 	}
 });
 
+const { ORIGIN_DEEPLINK, ORIGIN_QR_CODE } = AppConstants.DEEPLINKS;
+
 /**
  * PureComponent that manages ERC20 approve from the dapp browser
  */
@@ -297,7 +301,8 @@ class Approve extends PureComponent {
 		spendLimitCustomValue: undefined,
 		ticker: getTicker(this.props.ticker),
 		validSpendLimitCustomValue: true,
-		viewDetails: false
+		viewDetails: false,
+		spenderAddress: '0x...'
 	};
 
 	customSpendLimitInput = React.createRef();
@@ -312,13 +317,18 @@ class Approve extends PureComponent {
 		let tokenSymbol, tokenDecimals;
 		const contract = contractMap[safeToChecksumAddress(to)];
 		if (!contract) {
-			tokenSymbol = await AssetsContractController.getAssetSymbol(to);
-			tokenDecimals = await AssetsContractController.getTokenDecimals(to);
+			try {
+				tokenSymbol = await AssetsContractController.getAssetSymbol(to);
+				tokenDecimals = await AssetsContractController.getTokenDecimals(to);
+			} catch (e) {
+				tokenSymbol = 'ERC20 Token';
+				tokenDecimals = 18;
+			}
 		} else {
 			tokenSymbol = contract.symbol;
 			tokenDecimals = contract.decimals;
 		}
-		const originalApproveAmount = decodeTransferData('transfer', data)[2];
+		const [spenderAddress, , originalApproveAmount] = decodeTransferData('transfer', data);
 		const approveAmount = fromTokenMinimalUnit(hexToBN(originalApproveAmount), tokenDecimals);
 		const totalGas = gas.mul(gasPrice);
 		const { name: method } = await getMethodData(data);
@@ -330,7 +340,8 @@ class Approve extends PureComponent {
 			tokenDecimals,
 			tokenSymbol,
 			totalGas: renderFromWei(totalGas),
-			totalGasFiat: weiToFiatNumber(totalGas, conversionRate)
+			totalGasFiat: weiToFiatNumber(totalGas, conversionRate),
+			spenderAddress
 		});
 	};
 
@@ -438,7 +449,7 @@ class Approve extends PureComponent {
 	renderTransactionReview = () => {
 		const { host, method, viewData, tokenSymbol } = this.state;
 		const {
-			transaction: { to, data }
+			transaction: { to, data, origin }
 		} = this.props;
 		const amount = decodeTransferData('transfer', data)[1];
 
@@ -448,7 +459,7 @@ class Approve extends PureComponent {
 				toggleViewData={this.toggleViewData}
 				copyContractAddress={this.copyContractAddress}
 				address={renderShortAddress(to)}
-				host={host}
+				host={origin || host}
 				allowance={amount}
 				tokenSymbol={tokenSymbol}
 				data={data}
@@ -664,6 +675,7 @@ class Approve extends PureComponent {
 			this.trackApproveEvent(ANALYTICS_EVENT_OPTS.DAPP_APPROVE_SCREEN_APPROVE);
 		} catch (error) {
 			Alert.alert(strings('transactions.transaction_error'), error && error.message, [{ text: 'OK' }]);
+			Logger.error(error, 'error while trying to send transaction (Approve)');
 			this.setState({ transactionHandled: false });
 		}
 	};
@@ -696,13 +708,18 @@ class Approve extends PureComponent {
 			customGasVisible,
 			editPermissionVisible,
 			ticker,
-			gasError
+			gasError,
+			spenderAddress
 		} = this.state;
+
+		const {
+			transaction: { origin }
+		} = this.props;
 
 		const isFiat = primaryCurrency.toLowerCase() === 'fiat';
 		const currencySymbol = currencySymbols[currentCurrency];
 		const totalGasFiatRounded = Math.round(totalGasFiat * 100) / 100;
-
+		const originIsDeeplink = origin === ORIGIN_DEEPLINK || origin === ORIGIN_QR_CODE;
 		return (
 			<ActionModal
 				modalVisible={this.props.modalVisible}
@@ -730,12 +747,23 @@ class Approve extends PureComponent {
 					) : (
 						<>
 							<View style={styles.section} testID={'approve-screen'}>
-								<TransactionHeader currentPageInformation={{ title: host, url: activeTabUrl }} />
+								<TransactionHeader
+									currentPageInformation={{ origin, spenderAddress, title: host, url: activeTabUrl }}
+								/>
 								<Text style={styles.title} testID={'allow-access'}>
-									{strings('spend_limit_edition.allow_to_access', { tokenSymbol })}
+									{strings(
+										`spend_limit_edition.${
+											originIsDeeplink ? 'allow_to_address_access' : 'allow_to_access'
+										}`,
+										{ tokenSymbol }
+									)}
 								</Text>
 								<Text style={styles.explanation}>
-									{strings('spend_limit_edition.you_trust_this_site')}
+									{`${strings(
+										`spend_limit_edition.${
+											originIsDeeplink ? 'you_trust_this_address' : 'you_trust_this_site'
+										}`
+									)}`}
 								</Text>
 								<TouchableOpacity style={styles.actionTouchable} onPress={this.toggleEditPermission}>
 									<Text style={styles.editPermissionText}>