+++ app/components/Views/ChoosePasswordSimple/index.js
@@ -22,6 +22,7 @@ import { strings } from '../../../../locales/i18n';
 import { getNavigationOptionsTitle } from '../../UI/Navbar';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import zxcvbn from 'zxcvbn';
+import { getPasswordStrengthWord, passwordRequirementsMet } from '../../../util/password';
 
 const styles = StyleSheet.create({
 	mainWrapper: {
@@ -159,7 +160,7 @@ class ChoosePasswordSimple extends PureComponent {
 	onPressCreate = async () => {
 		if (this.state.loading) return;
 		let error = null;
-		if (this.state.password.length < 8) {
+		if (!passwordRequirementsMet(this.state.password)) {
 			error = strings('choose_password.password_length_error');
 		} else if (this.state.password !== this.state.confirmPassword) {
 			error = strings('choose_password.password_dont_match');
@@ -203,24 +204,6 @@ class ChoosePasswordSimple extends PureComponent {
 		}).start();
 	};
 
-	getPasswordStrengthWord() {
-		// this.state.passwordStrength is calculated by zxcvbn
-		// which returns a score based on "entropy to crack time"
-		// 0 is the weakest, 4 the strongest
-		switch (this.state.passwordStrength) {
-			case 0:
-				return 'weak';
-			case 1:
-				return 'weak';
-			case 2:
-				return 'weak';
-			case 3:
-				return 'good';
-			case 4:
-				return 'strong';
-		}
-	}
-
 	onPasswordChange = val => {
 		const passInfo = zxcvbn(val);
 
@@ -232,7 +215,8 @@ class ChoosePasswordSimple extends PureComponent {
 	};
 
 	render() {
-		const { startX, startY, width, height, initialScale, endX, endY } = this;
+		const { passwordStrength, startX, startY, width, height, initialScale, endX, endY } = this;
+		const passwordStrengthWord = getPasswordStrengthWord(passwordStrength);
 
 		return (
 			<SafeAreaView style={styles.mainWrapper}>
@@ -300,9 +284,9 @@ class ChoosePasswordSimple extends PureComponent {
 								{(this.state.password !== '' && (
 									<Text style={styles.passwordStrengthLabel}>
 										{strings('choose_password.password_strength')}
-										<Text style={styles[`strength_${this.getPasswordStrengthWord()}`]}>
+										<Text style={styles[`strength_${passwordStrengthWord}`]}>
 											{' '}
-											{strings(`choose_password.strength_${this.getPasswordStrengthWord()}`)}
+											{strings(`choose_password.strength_${passwordStrengthWord}`)}
 										</Text>
 									</Text>
 								)) || <Text style={styles.passwordStrengthLabel} />}