+++ app/components/Views/Entry/index.js
@@ -13,9 +13,9 @@ import { colors } from '../../../styles/common';
 import DeeplinkManager from '../../../core/DeeplinkManager';
 import Logger from '../../../util/Logger';
 import Device from '../../../util/Device';
-import SplashScreen from 'react-native-splash-screen';
 import { recreateVaultWithSamePassword } from '../../../core/Vault';
 import { EXISTING_USER, ONBOARDING_WIZARD, METRICS_OPT_IN, ENCRYPTION_LIB, ORIGINAL } from '../../../constants/storage';
+import AppConstants from '../../../core/AppConstants';
 
 /**
  * Entry Screen that decides which screen to show
@@ -96,7 +96,6 @@ class Entry extends PureComponent {
 	async componentDidMount() {
 		DeeplinkManager.init(this.props.navigation);
 		this.unsubscribeFromBranch = Branch.subscribe(this.handleDeeplinks);
-		SplashScreen.hide();
 		const existingUser = await AsyncStorage.getItem(EXISTING_USER);
 		if (existingUser !== null) {
 			await this.unlockKeychain();
@@ -114,7 +113,9 @@ class Entry extends PureComponent {
 		if (deeplink) {
 			const { KeyringController } = Engine.context;
 			const isUnlocked = KeyringController.isUnlocked();
-			isUnlocked ? DeeplinkManager.parse(deeplink) : DeeplinkManager.setDeeplink(deeplink);
+			isUnlocked
+				? DeeplinkManager.parse(deeplink, { origin: AppConstants.DEEPLINKS.ORIGIN_DEEPLINK })
+				: DeeplinkManager.setDeeplink(deeplink);
 		}
 	};
 
@@ -154,7 +155,7 @@ class Entry extends PureComponent {
 					NavigationActions.navigate({ routeName: 'Oboarding' })
 				);
 			} else {
-				this.props.navigation.navigate('HomeNav', {}, NavigationActions.navigate({ routeName: 'WalletView' }));
+				this.props.navigation.navigate('HomeNav');
 			}
 		});
 	};