+++ app/components/Views/ImportFromSeed/index.js
@@ -30,6 +30,7 @@ import TermsAndConditions from '../TermsAndConditions';
 import zxcvbn from 'zxcvbn';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import Device from '../../../util/Device';
+import { failedSeedPhraseRequirements } from '../../../util/validators';
 import { OutlinedTextField } from 'react-native-material-textfield';
 import {
 	BIOMETRY_CHOICE,
@@ -41,6 +42,11 @@ import {
 	METRICS_OPT_IN,
 	TRUE
 } from '../../../constants/storage';
+import { ethers } from 'ethers';
+import Logger from '../../../util/Logger';
+import { getPasswordStrengthWord, passwordRequirementsMet } from '../../../util/password';
+
+const { isValidMnemonic } = ethers.utils;
 
 const styles = StyleSheet.create({
 	mainWrapper: {
@@ -63,6 +69,16 @@ const styles = StyleSheet.create({
 	field: {
 		marginVertical: 5
 	},
+	fieldRow: {
+		flexDirection: 'row',
+		alignItems: 'flex-end'
+	},
+	fieldCol: {
+		width: '50%'
+	},
+	fieldColRight: {
+		flexDirection: 'row-reverse'
+	},
 	label: {
 		fontSize: 14,
 		marginBottom: 12,
@@ -77,7 +93,7 @@ const styles = StyleSheet.create({
 		...fontStyles.normal
 	},
 	seedPhrase: {
-		marginVertical: 10,
+		marginBottom: 10,
 		paddingTop: 20,
 		paddingBottom: 20,
 		paddingHorizontal: 20,
@@ -89,6 +105,9 @@ const styles = StyleSheet.create({
 		borderColor: colors.grey500,
 		backgroundColor: colors.white,
 		...fontStyles.normal
+	},
+	padding: {
+		paddingRight: 46
 	},
 	biometrics: {
 		alignItems: 'flex-start',
@@ -127,26 +146,21 @@ const styles = StyleSheet.create({
 	strength_strong: {
 		color: colors.green300
 	},
-	showHideToggle: {
-		backgroundColor: colors.white,
-		marginTop: 8,
-		alignSelf: 'flex-end'
-	},
 	showMatchingPasswords: {
 		position: 'absolute',
 		marginTop: 8,
 		alignSelf: 'flex-end'
 	},
 	qrCode: {
-		marginTop: -50,
-		marginBottom: 30,
-		alignSelf: 'flex-end',
 		marginRight: 10,
 		borderWidth: 1,
 		borderRadius: 6,
 		borderColor: colors.grey100,
 		paddingVertical: 4,
-		paddingHorizontal: 6
+		paddingHorizontal: 6,
+		marginTop: -50,
+		marginBottom: 30,
+		alignSelf: 'flex-end'
 	},
 	inputFocused: {
 		borderColor: colors.blue,
@@ -200,7 +214,8 @@ class ImportFromSeed extends PureComponent {
 		loading: false,
 		error: null,
 		seedphraseInputFocused: false,
-		inputWidth: Device.isAndroid() ? '99%' : undefined
+		inputWidth: { width: '99%' },
+		hideSeedPhraseInput: true
 	};
 
 	passwordInput = React.createRef();
@@ -216,29 +231,27 @@ class ImportFromSeed extends PureComponent {
 			}
 			this.setState({ biometryType, biometryChoice: enabled });
 		}
-		this.mounted = true;
 		// Workaround https://github.com/facebook/react-native/issues/9958
-		this.state.inputWidth &&
-			setTimeout(() => {
-				this.mounted && this.setState({ inputWidth: '100%' });
-			}, 100);
+		setTimeout(() => {
+			this.setState({ inputWidth: { width: '100%' } });
+		}, 100);
 	}
 
-	componentWillUnmount = () => {
-		this.mounted = false;
-	};
-
 	onPressImport = async () => {
-		if (this.state.loading) return;
+		const { loading, seed, password, confirmPassword } = this.state;
+
+		if (loading) return;
 		let error = null;
-		if (this.state.password.length < 8) {
+		if (!passwordRequirementsMet(this.state.password)) {
 			error = strings('import_from_seed.password_length_error');
-		} else if (this.state.password !== this.state.confirmPassword) {
+		} else if (password !== confirmPassword) {
 			error = strings('import_from_seed.password_dont_match');
 		}
 
-		if (this.state.seed.split(' ').length !== 12) {
-			error = strings('import_from_seed.seed_word_count_error');
+		if (failedSeedPhraseRequirements(seed)) {
+			error = strings('import_from_seed.seed_phrase_requirements');
+		} else if (!isValidMnemonic(seed)) {
+			error = strings('import_from_seed.invalid_seed_phrase');
 		}
 
 		if (error) {
@@ -311,6 +324,7 @@ class ImportFromSeed extends PureComponent {
 					this.setState({ loading: false });
 				} else {
 					this.setState({ loading: false, error: error.toString() });
+					Logger.log('Error with seed phrase import', error);
 				}
 			}
 		}
@@ -389,35 +403,26 @@ class ImportFromSeed extends PureComponent {
 		this.setState({ secureTextEntry: !this.state.secureTextEntry });
 	};
 
-	getPasswordStrengthWord() {
-		// this.state.passwordStrength is calculated by zxcvbn
-		// which returns a score based on "entropy to crack time"
-		// 0 is the weakest, 4 the strongest
-		switch (this.state.passwordStrength) {
-			case 0:
-				return 'weak';
-			case 1:
-				return 'weak';
-			case 2:
-				return 'weak';
-			case 3:
-				return 'good';
-			case 4:
-				return 'strong';
-		}
-	}
+	toggleHideSeedPhraseInput = () => {
+		this.setState(({ hideSeedPhraseInput }) => ({ hideSeedPhraseInput: !hideSeedPhraseInput }));
+	};
 
 	onQrCodePress = () => {
+		setTimeout(this.toggleHideSeedPhraseInput, 100);
 		this.props.navigation.navigate('QRScanner', {
-			onScanSuccess: meta => {
-				if (meta && meta.seed) {
-					this.setState({ seed: meta.seed });
+			onScanSuccess: ({ seed = undefined }) => {
+				if (seed) {
+					this.setState({ seed });
 				} else {
 					Alert.alert(
 						strings('import_from_seed.invalid_qr_code_title'),
 						strings('import_from_seed.invalid_qr_code_message')
 					);
 				}
+				this.toggleHideSeedPhraseInput();
+			},
+			onScanError: error => {
+				this.toggleHideSeedPhraseInput();
 			}
 		});
 	};
@@ -427,48 +432,92 @@ class ImportFromSeed extends PureComponent {
 	render() {
 		const {
 			password,
+			passwordStrength,
 			confirmPassword,
 			seed,
 			seedphraseInputFocused,
 			inputWidth,
 			secureTextEntry,
 			error,
-			loading
+			loading,
+			hideSeedPhraseInput
 		} = this.state;
+
+		const passwordStrengthWord = getPasswordStrengthWord(passwordStrength);
 
 		return (
 			<SafeAreaView style={styles.mainWrapper}>
 				<KeyboardAwareScrollView style={styles.wrapper} resetScrollToCoords={{ x: 0, y: 0 }}>
 					<View testID={'import-from-seed-screen'}>
 						<Text style={styles.title}>{strings('import_from_seed.title')}</Text>
-						<TextInput
-							value={seed}
-							numberOfLines={3}
-							multiline
-							style={[
-								styles.seedPhrase,
-								inputWidth && { width: inputWidth },
-								seedphraseInputFocused && styles.inputFocused
-							]}
-							placeholder={strings('import_from_seed.seed_phrase_placeholder')}
-							placeholderTextColor={colors.grey200}
-							onChangeText={this.onSeedWordsChange}
-							testID={'input-seed-phrase'}
-							blurOnSubmit
-							onSubmitEditing={this.jumpToPassword}
-							returnKeyType={'next'}
-							keyboardType={Device.isAndroid() ? 'visible-password' : 'default'}
-							autoCapitalize="none"
-							autoCorrect={false}
-							onFocus={this.seedphraseInputFocused}
-							onBlur={this.seedphraseInputFocused}
-						/>
+						<View style={styles.fieldRow}>
+							<View style={styles.fieldCol}>
+								<Text style={styles.label}>{strings('choose_password.seed_phrase')}</Text>
+							</View>
+							<View style={[styles.fieldCol, styles.fieldColRight]}>
+								<TouchableOpacity onPress={this.toggleHideSeedPhraseInput}>
+									<Text style={styles.label}>
+										{strings(`choose_password.${hideSeedPhraseInput ? 'show' : 'hide'}`)}
+									</Text>
+								</TouchableOpacity>
+							</View>
+						</View>
+						{hideSeedPhraseInput ? (
+							<OutlinedTextField
+								containerStyle={inputWidth}
+								inputContainerStyle={styles.padding}
+								placeholder={strings('import_from_seed.seed_phrase_placeholder')}
+								testID="input-seed-phrase"
+								returnKeyType="next"
+								autoCapitalize="none"
+								secureTextEntry={hideSeedPhraseInput}
+								onChangeText={this.onSeedWordsChange}
+								value={seed}
+								baseColor={colors.grey500}
+								tintColor={colors.blue}
+								onSubmitEditing={this.jumpToPassword}
+							/>
+						) : (
+							<TextInput
+								value={seed}
+								numberOfLines={3}
+								style={[styles.seedPhrase, inputWidth, seedphraseInputFocused && styles.inputFocused]}
+								secureTextEntry
+								multiline={!hideSeedPhraseInput}
+								placeholder={strings('import_from_seed.seed_phrase_placeholder')}
+								placeholderTextColor={colors.grey200}
+								onChangeText={this.onSeedWordsChange}
+								testID="input-seed-phrase"
+								blurOnSubmit
+								onSubmitEditing={this.jumpToPassword}
+								returnKeyType="next"
+								keyboardType={
+									(!hideSeedPhraseInput && Device.isAndroid() && 'visible-password') || 'default'
+								}
+								autoCapitalize="none"
+								autoCorrect={false}
+								onFocus={(!hideSeedPhraseInput && this.seedphraseInputFocused) || null}
+								onBlur={(!hideSeedPhraseInput && this.seedphraseInputFocused) || null}
+							/>
+						)}
 						<TouchableOpacity style={styles.qrCode} onPress={this.onQrCodePress}>
 							<Icon name="qrcode" size={20} color={colors.fontSecondary} />
 						</TouchableOpacity>
 						<View style={styles.field}>
-							<Text style={styles.label}>{strings('import_from_seed.new_password')}</Text>
+							<View style={styles.fieldRow}>
+								<View style={styles.fieldCol}>
+									<Text style={styles.label}>{strings('import_from_seed.new_password')}</Text>
+								</View>
+								<View style={[styles.fieldCol, styles.fieldColRight]}>
+									<TouchableOpacity onPress={this.toggleShowHide}>
+										<Text style={styles.label}>
+											{strings(`choose_password.${secureTextEntry ? 'show' : 'hide'}`)}
+										</Text>
+									</TouchableOpacity>
+								</View>
+							</View>
 							<OutlinedTextField
+								containerStyle={inputWidth}
 								ref={this.passwordInput}
 								placeholder={strings('import_from_seed.new_password')}
 								testID={'input-password-field'}
@@ -480,21 +529,14 @@ class ImportFromSeed extends PureComponent {
 								baseColor={colors.grey500}
 								tintColor={colors.blue}
 								onSubmitEditing={this.jumpToConfirmPassword}
-								renderRightAccessory={() => (
-									<TouchableOpacity onPress={this.toggleShowHide} style={styles.showHideToggle}>
-										<Text style={styles.passwordStrengthLabel}>
-											{strings(`choose_password.${secureTextEntry ? 'show' : 'hide'}`)}
-										</Text>
-									</TouchableOpacity>
-								)}
 							/>
 
 							{(password !== '' && (
 								<Text style={styles.passwordStrengthLabel}>
 									{strings('choose_password.password_strength')}
-									<Text style={styles[`strength_${this.getPasswordStrengthWord()}`]}>
+									<Text style={styles[`strength_${passwordStrengthWord}`]}>
 										{' '}
-										{strings(`choose_password.strength_${this.getPasswordStrengthWord()}`)}
+										{strings(`choose_password.strength_${passwordStrengthWord}`)}
 									</Text>
 								</Text>
 							)) || <Text style={styles.passwordStrengthLabel} />}
@@ -503,6 +545,7 @@ class ImportFromSeed extends PureComponent {
 						<View style={styles.field}>
 							<Text style={styles.label}>{strings('import_from_seed.confirm_password')}</Text>
 							<OutlinedTextField
+								containerStyle={inputWidth}
 								ref={this.confirmPasswordInput}
 								testID={'input-password-field-confirm'}
 								onChangeText={this.onPasswordConfirmChange}
@@ -540,7 +583,11 @@ class ImportFromSeed extends PureComponent {
 								onPress={this.onPressImport}
 								testID={'submit'}
 								disabled={
-									!(password !== '' && password === confirmPassword && seed.split(' ').length === 12)
+									!(
+										password !== '' &&
+										password === confirmPassword &&
+										!failedSeedPhraseRequirements(seed)
+									)
 								}
 							>
 								{loading ? (